+++ src/Notices/OnCommand.cs
@@ -1,4 +1,5 @@
-﻿using HarmonyLib;
+﻿using System.Collections.Generic;
+using HarmonyLib;
 using JetBrains.Annotations;
 
 namespace DiscordBot.Notices;
@@ -12,7 +13,20 @@ public static class OnCommand
         private static void Postfix(Terminal.ConsoleCommand __instance, Terminal.ConsoleEventArgs args)
         {
             if (!DiscordBotPlugin.ShowCommandUse || !__instance.IsCheat || __instance.IsPluginCommand()) return;
-            Discord.instance?.SendMessage(Webhook.Notifications, Player.m_localPlayer?.GetPlayerName() ?? "Unknown", $"Executed command: `{string.Join(" ", args.Args)}`", DiscordBotPlugin.OnUseCommandHooks);
+            if (Player.m_localPlayer)
+            {
+                Dictionary<string, string> details = new Dictionary<string, string>()
+                {
+                    ["Coordinates"] = $"{Player.m_localPlayer.transform.position.x:0.0}, {Player.m_localPlayer.transform.position.y:0.0}, {Player.m_localPlayer.transform.position.z:0.0}",
+                    ["Biome"] = Player.m_localPlayer.GetCurrentBiome().ToString(),
+                };
+                
+                Discord.instance?.SendTableEmbed(Webhook.Notifications, $"Executed command: `{string.Join(" ", args.Args)}`", details, Player.m_localPlayer.GetPlayerName(), hooks: DiscordBotPlugin.OnUseCommandHooks);
+            }
+            else
+            {
+                Discord.instance?.SendMessage(Webhook.Notifications, ZNet.instance.GetWorldName(), $"Executed command: `{string.Join(" ", args.Args)}`", DiscordBotPlugin.OnUseCommandHooks);
+            }
         }
     }
 }